PRIMERO-293 Date Bug - Date picker will not set the selected date.
date_controls don't use getDate method, it will returns the current date if datepicker value is not valid,
there is no way to know that the current value is valid, so parse again the value in order to format to
the expected value. parseDate try to several formats to decide this is a valid value.

Create initialize_datepicker in order to call in other places and have the commons properties.

age_birth_date don't rely on the focus event to initialize the datepicker if is not already a datepicker, call
initialize_datepicker to get properly initialized the control in order to get the expected format. Also this fix
the issue when edit age and have to click two times the save button.

The test case fixed should pass so enter a valid date, before the date control change the value to the current date,
so know put the correct value.