PRIMERO-358 display designated values when collapse/expanding subforms when Editing - Family Details - CP-039/047/024. Refactoring subform header: it will show values separed by dash character without the corresponding label field.
Refactoring the way that update the header: now the fields that belong to the header are send in the attributes data-types and data-fields,
the values are lookup in the page and they are printed separated by dash character. If there is no value for a field, this is not diplayed.

Refactoring step that verify the subform header.