PRIMERO-607 display text for nested subforms in view mode. 1. The following text will be displayed when the 'Photos and Audio' section is empty: Click the EDIT button to add Photos and Audio details. 2. The following text will be displayed when the 'Other Documents' section is empty: Click the EDIT button to add Other Documents. 3. In FormSection added the property 'display_help_text_view' that indicate if we are going to display the messages when the section is empty.
This change will apply to 'Other Documents' on Incident and Cases.
This change will apply to 'Photos and Audio' on Cases and Tracing Request.

The messages can be apply to Sections that has one field which is a subform in order to display the message if the subform is empty.