PRIMERO-537 Clean up logs generated by the system.
Make the logs for the different services be located in the same directory,
by default all go to log directory (subdirectories nginx, solr, rails and couchdb).

The log for rails is defined by the environment variable "RAILS_LOG_PATH", the file name is
take from the environment variable RAILS_ENV, it can be production, development, etc.

Added monkey patch to be able to move the sunspot/solr log, the path file can be indicated
in the sunspot.yml file via 'log_file', but this variable is read by sunspot but it is ignored
when the server is being configured, so added monkey patch to make it happen the read.