PRIMERO-798 Date Validation on Date Range Fields.
Validate the date range when is just a single date. Ask the 'subtype' date or date_range to identify the corresponding field to validate.
Cleanup the values when the user switch between date and date_range so possible invalid values are not send to the controller and cleanup the old values.
Refactoring test cases.