PRIMERO-823 UAT17: link to GBV incidents.
Adding incident_links custom field. The field can't be delete. The database name can't be change.
The custom field define the property 'custom_template' which point to the template to render the field.
incident_links is populated when the user create a GBV incident from a GBV case.
incident_links belong to the action plan form section.