SL-18 Age Range Filter is not an "and" filter.
The issue is that range filter was taking only the first of the range selected
for example when the filter send was 6-11||12-17, the logic take in count only
6-11.

The logic was fixed on the parameters processing and on the sunspot filter build.

Also on Incident range age filter was not well defined, fixed as well.