SL-244 Prevent users from changing field types.
db/forms/case/survivor_information.rb and db/forms/case/basic_identity.rb define the "sex" field
with different type, fix the type to match.

Fix error messages, was printing the Form Section the user is changing or updating not the
Form Section were was found the other field with different type.

Fix the list of field over what validate, it was including fields in subforms, so if several
subforms contains fields with the same name and different types it throw the validation, but
that is not right, subforms fields with the same and different types should be possible.

Added more test cases.