SL-536 QA2: nested form not editable.
Workaround in the matching of routes to identify which javascript code should be
executed by page. Backbone router is used to match which javascript code belong to the page.

When the user save the record and some error occurrs, the URL in the navigation bar change
from cases/ff72ddc52b33be0650f91071207c5eb0/edit?follow=true to cases/ff72ddc52b33be0650f91071207c5eb0
and this confuse Backbone and load the javascript supposed to be on the view page and the edit page as well.

That make the subforms events got registered twice and when the user expand the subform is execute twice
and collapse after expand and look like is doing nothing.

This is mostly a workaround.