JOR-84 Pending Approvals Filter.
build_filter_checkboxes_group was introduced because the current implementation
does not allow have different fields name for the same Filter Title.

In the case of this filter, the title is the same 'Pending Approvals' but the checboxes
should point to 3 different fields names, in this way we don't need to make anything else
in the controller, just send the field with the corresponding value to search.