JOR-142 Accept/Reject Transfers.
Display the Accept and Reject button on the transition subform on view page.
Display the Accept and Reject dialog box.
Children controller Accept and Reject action to process the record.

When the user Accept the case, became the record owner and the Accept/Reject button disappears for the transfer.
When the user Reject the case, the case is not longer accessable to this user and  the Accept/Reject button disappears for the transfer.

In progress.