PRIM-913 - Locations not loading on add-service-provision modal. Updated cached elements.
After a search the refer/transfer modals disappear from the action menu,
thus, there are no ReportingLocation elements. When a user clicks the
add service provision button, the script does nothing because
no elements are found. Now, the script that setups a single select also
updates the elements found in the page.