PRIM-957 - Manager tasks dashboard has inconsistent numbers
The dashboard of overdue tasks is incorrectly showing users
as owners of cases they were referred. This causes inconsistencies
between the case list and the dashboard, for example, a user may appear with
number of cases that he doesn't own, clicking the number will show a different
number of cases in the case list. Now we only show the case owners in the dashboard.