Remove the duplicate "hide_name" key in "en.yml"
The instance of "hide_name" that appears to be actually used is
structured like this:

    hide_name:
      protect: "Hide Name"
      view: "View Name"

I grepped through the code and found that "hide_name" only appears
when one of the keys under it ("protect" or "view") is accessed.

Thus, I elected to keep this instance and remove the other one.