Remove the duplicate "high_risk" key in "en.yml"
There is no good answer for this one.  It is completely unclear
which one of the duplicates should be used.

I removed the selected duplicate because "High Priority" matches the
wording of the other "*_risk" keys in the same block.

This probably deserves review by someone who knows more than me.