Define a function to build list of preserved keys
This function is a bit complicated.

First, we convert our key set into a list of strings that is sorted
by string length.  This is necessary so that key prefixes will have
priority over their sub-keys.  For example, if some location in the
code uses "logging.#{foo}" as a key, then we need to preserve the
whole logging group, even if a more specific sub-key of "logging" is
used later.

Sorting by length allows us to maintain this preference because any
string sub-key of "logging" will be longer!