Check for multiple translations per line
This is confusing, so I'll add some detail...

Initially, I thought I had covered the case of multiple matches
because I used the "-o" option in my grep command.  The man page for
grep reads:

  -o, --only-matching
   Print only the matched (non-empty) parts of a matching line, with
   each such part on a separate output line.

I noticed, still, that multiple matches were showing up in a single
result, so I started to investigate.

The command below does what I expect with the "-o" option:

  $ echo 'foo foo' | grep -on 'foo'
  1:foo
  1:foo

Notice that we get two results for each match on our input line.

However, when we change the pattern to "foo.*", we get a surprise.

  $ echo 'foo foo' | command grep -on 'foo.*'
  1:foo foo

This is because grep moves on once it has consumed all of the
characters to satisfy the pattern.  The output I had expected was:

  1:foo foo
  1:foo

---

The end result is that I had to add support for checking for
multiple matches in a single grep result, even though I'm already
trimming the output somewhat with the "-o" option.

Basically, I take the result text, and split it into a list of text
snippets in between matches to a regular expression that is
equivalent to the one I was searching for before.

It may add a nested "for" loop, but it should catch everything!