Limit header typography rules for CardHeader
Previously any span within a CardHeader element would be styled
with the typography rules. This is fine when an OptionBoxes action
never contains any text but when it does it causes all of the text
to have the same styles. This reflects poorly the true information
heriarchy of the title.

This limits the rules to only the typography of the title of the
OptionBox. Unfortunately I've had to move those styles into the
element definition. Maintaining two ways of specifying styles for
this element might be a bit much over head. If there is a way to
facilitate this with css-to-mui-loader.js then we should do that
but it appears not possible without significant adaptations to
css-to-mui-loader.js which seems a bit out of scope for this.