Add Number of Incidents KPI UI
I'll likely need to move to programatically grabbing the KPIs
from the backend but for now this is sufficient and avoids
attempting to build a bad abstraction over all the different ways
to display KPIs.

I've also replaced the previous date picker with a simpler
select. I still need to implement a dialog to capture the
custom date ranges from the user though.