Add Assessment Status KPI
I've implemented a new component for this, the
StackedPercentageBar. This will need to be factored out of
the KPIs component at a later date.

I'm not totally happy with the html / css architecture at the
moment. I suspect I will like to move things around abit.

We should also add aria labels to this stuff.