Fix vertical alignement for TablePercentageBars
inline-block elements seem to to be aligned to the base of the line
but the element for the percentage bar needs to be centered in the
partent element. This is achieved here by aligning the element to
the top of the line (same behaviour as block elements) as it is
the same height as the parent. It might have been better to align
center.