Move to I18n library for localizing dates
Unfortuantely there seems to be a bug wrt to using 'useI18n'.
When called from within this element it returns a null object.
I've not been able to track down the issue so I'm using the global
I18n object available at 'window.I18n'. This is the same object that
is returned but isn't the ideal way to retieve access to it.