PRIM-1677 Legacy Python2 issues.
Install python3 only and create symbolic links to python2 and pip of python2
that point to python3 to make looks like python2/pip is installed in order
to avoid fails of the cookbook of supervisor and python.

For for reason in the case of solr is not stopping randomly solr
can we just kill solr?