PRIM-1700 - API - "Format" parameter causing bulk exports to show only json
Format is a parameter used by rails to determine the response format.
This param was overwriting the format param we use to filter the bulk exports in the index action.

The param was renamed to export_format using an alias_attribute so
that we don't need to change the db.

A small fix for sorting the index results was added.