PRIM-1677 Legacy Python2 issues.
'supervisorctl restart solr' isn't working because it seems like solr
take sometimes in obey the termination signal and when run the start the
process still there and so it fails.

Just calling Stop and Start it doesn't work neither, so we relay
on the sleep method at the OS to make time to the process obey the signal.

Tried stopwaitsecs of supervisor but didn't work.