Remove minimum_count on pivot
A minimum_count on the pivot will return all of the counts over the
range for the chose pivot field. This would mean we have all of the
locations with lots of 0's if those locations had no <insert thing> done
at them. This would ruin the UI. We only need show locations we have
data for.