Add query for ReportingDelay
Also I've added fixture data but there should be a file specific to this
KPI.

In this commit I make use of the adjust_solr_params function from
Sunspot. This is to the the frange feature of queries in solr. Sunspot
currently doesn't have this feature so I will explore patching it into
Sunpot later.