PRIM-1794 - Avoid n + 1 issue when fetching reporting_locations for users
When doing referrals we need to filter users by reporting location and
since there's no relationship between location and users, we need to perform
a query against location for each user.

A service was created to query all the locations for users and sets them
in the user_location and reporting_location fields.