Address Feedback 19: The Date Range selector has been upgraded.
Previously it would only work when selecting from a not custom-range
date range as no change event is fired when selecting and alreadys
elected item. Now it makes use of the custom-range MenuItem click event
which means that the date range select dialog will always be shown when
you click on the custom-range MenuItem.