R2-1024 - Clicking into the service type field clears out service fields
Use a local state to hold the initial values and only recalculate them
when the dialog is open instead of passing them as props, because the default values
for dates, e.g. "now" will be recalculated on each render resulting in a different
value that will cause a reset of the form.