R2-1079 - Do not lose the selectedRecord when records are updated after going back online
When offline, a user can perform certain actions that might take
a record out of the records list once he gets back online.

For instance, when offline a user might disable a record,
go to the records list and then back to the disabled record.
Once online all those requests are performed, resulting in a
record not found error because the disabled record
is not longer in the records list.

To prevent this error, the selectedRecord is now stored under the
specific recordType, so that, when the data is updated,
the selectedRecord is updated instead of deleted.