R2-1242 - Multiselect are not updated if all the selected values are removed
When a multiselect was cleared, the values were not being updated in
the server, because the new value was an empty array and it was being discarded
from the submitted values.

Fixes in compactValues function:

* Do not discard empty array as new values if the previous array had elements.
* Do not compare undefined keys when a subform is found. That is to say,
if a key "id" is present in the old value, the new value must have an "id" key.