Perf improvements for kpi calculations
We're reducing the number of queries needed in the pathalogical case
when calculating kpi data. The kpis often rely on form and fields which
can form a tree of arbitrary depth. By eager loading the first 2 levels
(the current form, it's fields, any subforms and their fields) we can
avoid most of the extra queries.

This may need to be adapted if forms become more deeply nested though I
don't expect this to be encouraged.