R2-1366 - Remove alerts only if an update is performed through the API.
Alerts are removed when a record is updated by its owner, however,
there are scenarios where alerts might need to be preserved, for instance,
when an update is performed through a migration script or when
an incident creates an alert for a case.

A new attribute record_user_update was created to indicate that an update on
a record was performed by an API call and therefore, alerts can be safely
removed.