fix: faceted date range queries for kpis
Previously we were using client tz for date calculations on the clients
side, converting them to UTC when sending to the server and then using
facet.interval.gap +1MONTH for Solar queries.

This means that:
1. Client times are often out by a number of hours e.g. beginning of the
months in EST becomes 19:00 on the last day of the previous month.
2. Lack of facet.interval.hardend meant that Solr attempted to fill the
last gap in the interval. If the last interval were not a whole month
then SOlr might include some of the next month in the query.

These two things seem to interfer such that range queries in certain
timezones at certain times of the year returned the wrong set of dates.

By moving to using UTC to do date calulation on the frontend and
introducing facet.interval.hardend this should be mitigated.