show utc times for kpis as local times
this is because the times actually represent a month period but we use a
month boundary to represent this e.g. frst day of month at 00:00:00.
When converted into local time in a tz like -05:00 we get the last day
of the previous month at 08:00:00 which, when we render the month only,
will give us the previous month, which is totally wrong.