fix tallying of lookup values
we're using different lookups now. i wonder if there is a way to
automatically keep these in sync? we don't currently capture any
semantic value on the lookups themselves. if we were to tag them with a
concept indicating positive, negative etc then we would be able to pull
them form the config rather than updating these values with the config.