updates display of client_satisfaction_rate kpi
atm we're not really calculating a usefule measure of client
satisfaction. Until we are, small values of satisfaction are hidden
behind a message indicating that they are not useful. This is removed
when sufficient data exists to display a useful message.

it also hides situations where more client provide negative feedback
than positive feedback because it doesn't take negative feedback into
account!