Move to using UTC date for testing.
`date-fns:parseISO` parses the given string in the local timezone.
Converting a `Date` to string via `toISOString` converts a local
datetime into UTC datetime. Since `endOfDay` takes that UTC datetime,
manipulates it as a string and passes it off to `date-fns:parseISO`
again, it's going to be impacted by timezone conversion twice.

This is fine if the local timezone is UTC-n because T00:00:00 + 0n:00:00
(used to convert the local time to UTC time) is still the same date.

If the local timezone is UTC+n then it becomes T00:00:00 - 0n:00:00
which will push the date back into the previous day! So the string
output from `toISOString` in a UTC+n timezone will be of the previous day
(and month, year if the day is on a boundary).

Strings passed to `new Date` are assumed to be in UTC so when we use
that in the test, it works in UTC+n and UTC-n timezones!

Dates are hard.