R2-1891 - Fix query to pull permitted form
The permitted_forms query was not returning the subforms
when the include_subforms parameter is defined.

The base exporter was trying to pull the permitted_forms and
the permitted_subforms for a user, but the permitted_forms query
was inadvertently retrieving subforms if they were associated
to the role resulting in duplicated subforms, which in turn produced
duplicated tabs in the exported excel files.