R2-2455 - Do not duplicate subform tabs in excel exports
Since subforms were being pulled as part of the permitted forms,
that caused duplication of tabs, because a form will export its
subforms and at the same time the subforms are exported as separated
forms.

Now subforms are not included as permitted forms, instead they are
pulled from the subform fields when a form is exported.