R2-2487 - Changing date filter field does not clear out old date filter
In react hook form v6, when shouldUnregister: false the unregister
method does not longer works, that is, all the registered fields
are submitted. Here the date field values are removed when the selected
date field changes.

The local state that used to store the current selected filters is
also removed, because filters are pulled from the query params.