R2-2689 - Dont merge permitted_forms
Before a record is stored in indexeddb, the old data of the
record is merged with the new data, that was causing a merge of
the permitted forms, as a result, a user will always have access
to all forms that were permitted at some point, even though
new requests specify different permissions.

Now, we overwrite the permitted forms so that users will only
have access to those forms specified by the last request performed.