R2-2919 - Fix invalid json and offline duplication when a document is updated
When an update was performed against an attachment the field_name
was part of the request and that field can not be changed for
an existing attachment. For that reason the server returned an
invalid json response.

When a document is updated while offline, the update must not
create another document.