R2-3008 - Fix: Readable fields are not permitted when an update is performed
Once the fields attribute is set in the PermittedFormFieldsService,
it won't update unless a new instance is created or a force = true
parameter is passed. That means the same list of permitted fields will be
returned for writeable or readable fields when it happens in a single request.

Because the writeable attribute is part of the cache's key, the service will
now always hit the cache instead of returning those stored in the fields attribute.