R2-3489 - Click-through from case to linked family gives you a 404
Clicking a family link redirected the user to the family page,
but the selected form remained families_from_case.
This caused an invalid fetch to the cases API,
since the ID used was not valid for that request.

Now we reset the form navigation when a different record
is loaded.