R2-3490 - User told they cannot access case created from family
When a case was created for a family member, the link to access
the case was incorrectly disabled, making it appear as
if access was forbidden. This also affected other family members
with previously accessible cases.

Now the response includes the can_read_record property to correctly
indicate whether the user has access to each linked case.