Fix: Eror message when the page /my_case/new is refreshed
When the /my_case/new page is refreshed a fetch request
against /cases/identified is performed, if it returns 404
a server error message was displayed because no redirection
callback was triggered.

Now we always trigger a redirection callback since a redirection
against the same location is a no-op as it causes no additional
render.

Additionally we are reusing the dispatchIfStatus logic and
removed all the code related to the redirectOnNotFound property.